+++ packages/expo-cli/src/commands/eas-build/utils/expoUpdates/isExpoUpdatesInstalled.ts
@@ -1,0 +1,7 @@
+import { getPackageJson } from '@expo/config';
+
+export default function isExpoUpdatesInstalled(projectDir: string) {
+  const packageJson = getPackageJson(projectDir);
+
+  return packageJson.dependencies && 'expo-updates' in packageJson.dependencies;
+}