+++ packages/expo-cli/src/commands/eas-build/utils/git.ts
@@ -1,5 +1,7 @@
 import spawnAsync from '@expo/spawn-async';
+import chalk from 'chalk';
 import commandExists from 'command-exists';
+import figures from 'figures';
 import fs from 'fs-extra';
 import ora from 'ora';
 
@@ -110,10 +112,58 @@ async function reviewAndCommitChangesAsync(
   await spawnAsync('git', ['commit', '-m', message]);
 }
 
+async function modifyAndCommitAsync(
+  callback: () => Promise<void>,
+  {
+    startMessage,
+    successMessage,
+    commitMessage,
+    commitSuccessMessage,
+    nonInteractive,
+  }: {
+    startMessage: string;
+    successMessage: string;
+    commitMessage: string;
+    commitSuccessMessage: string;
+    nonInteractive: boolean;
+  }
+) {
+  const spinner = ora(startMessage);
+
+  try {
+    await callback();
+
+    await ensureGitStatusIsCleanAsync();
+
+    spinner.succeed();
+  } catch (err) {
+    if (err instanceof DirtyGitTreeError) {
+      spinner.succeed(successMessage);
+      log.newLine();
+
+      try {
+        await reviewAndCommitChangesAsync(commitMessage, {
+          nonInteractive,
+        });
+
+        log(`${chalk.green(figures.tick)} ${commitSuccessMessage}.`);
+      } catch (e) {
+        throw new Error(
+          "Aborting, run the command again once you're ready. Make sure to commit any changes you've made."
+        );
+      }
+    } else {
+      spinner.fail();
+      throw err;
+    }
+  }
+}
+
 export {
   DirtyGitTreeError,
   ensureGitRepoExistsAsync,
   ensureGitStatusIsCleanAsync,
   makeProjectTarballAsync,
   reviewAndCommitChangesAsync,
+  modifyAndCommitAsync,
 };